feat: 🎸 sqlite connection
Added an implementation to talk with/create an sqlite db file that is
set up to track the commit count meant for a specific commit plan. There
is also a users table in the database, but this will most likely need to
be changed in the future if it will be meant to made use of. Right now
it only has some fields that most likely won't really be used by the
commiter.

✅ Closes: 1